Fixed: -NPC spawning location problem -Object files Changes/Addition: -Renamed NPC* ptr[10] to NPCs[10] -Renamed sizeofarray to NPClimit -Added Entity* entities[11] and int entityLimit(size of this array) -Added occupied function in game.cpp to check what entity is in a location -Added virtual function type() to entites

+++ SP1Framework/Position.cpp
@@ -1,5 +1,5 @@
 #include "Position.h"
-Position::Position(int x, int y)
+Position::Position(float x, float y)
 {
 	this->x = x;
 	this->y = y;
